Change the way the test server operated by using a magic gateway URL.
The "test mode" was difficult to configure. This new method makes it part
of the gateway URL which avoids confusion and makes it easier to configure
in user code as the gateway URL value will already be a configuration item.

Also, the completation URL function now receives the original request
object and the full transaction. This is a breaking change.